refactor: change int conversion to float

+++ 10-Calculator/calculator.py
@@ -1,6 +1,4 @@
 # calculator
-
-
 # add
 def add(n1, n2):
     return n1 + n2
@@ -29,25 +27,29 @@ operations = {
     "/": divide,
 }
 
+
 def calculator():
-    num1 = int(input("What's the fist number? "))
+    num1 = float(input("What's the fist number? "))
     for symbol in operations:
         print(symbol)
 
     should_continue = True
     while should_continue:
         operation_symbol = input("Pick an operation: ")
-        num2 = int(input("What's the next number? "))
+        num2 = float(input("What's the next number? "))
 
         answer = operations[operation_symbol](num1, num2)
         print(f"{num1} {operation_symbol} {num2} = {answer}")
         if (
-            input(f"Press [y] to continue calculating with {answer} or type any key to start a new calculation: ")
+            input(
+                f"Press [y] to continue calculating with {answer} or type any key to start a new calculation: "
+            )
             == "y"
         ):
             num1 = answer
         else:
             should_continue = False
             calculator()
-            
+
+
 calculator()
